backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pandas
pydantic
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
python-multipart
google-generativeai
reportlab
python-dotenv
alembic
PyPDF2
